refactor(ContactTpl): port typescript
