+++ gatsby-node.js
@@ -14,7 +14,7 @@ exports.createPages = async ({ graphql, actions }) => {
   const definePageTpl = uid => {
     switch (uid) {
       case "contact":
-        return path.resolve("lib/ui/templates/ContactTpl.js");
+        return path.resolve("lib/ui/templates/ContactTpl.tsx");
       case "eotm":
         return path.resolve("lib/ui/templates/EmojiTpl.js");
       case "about":
